Updated Solution and added Mutation.

+++ GeneticCars/GeneticCars/GameLogic/GameRules.swift
@@ -6,12 +6,12 @@
 //  Copyright © 2017 Sebastian Kotarski. All rights reserved.
 //
 
-import Foundation
+import CoreGraphics
 
 struct GameRules {
     static let coordinatesRangeForHorizontal = 50
     static let coordinatesRangeForVertical = 50
-    static let wheelSizeRange = 20
+    static let wheelSizeRange = CGFloat(20)
     static let numberOfPointsInSolution = 6
     static let numberOfWheelsInSolution = 2
 }
